Gitignore fájl hozzáadva. Controller pdfre előkészítve, a pedf generáló részre vár, hogy tesztelhessük

+++ SzakdolgozatErtekeloApi/DTO/ErtekelesDTO.cs
@@ -3,7 +3,7 @@
     public class ErtekelesDTO
     {
         public int ID { get; set; }
-        public double Jegy { get; set; }
+        public float Jegy { get; set; }
         public string Megjegyzes { get; set; }
     }
 }
